refactor: Home 페이지 폴더 구조 리팩터링
