+++ src/pages/Home/components/NavBar.tsx
@@ -2,8 +2,8 @@
 import { css } from '@emotion/react';
 import { FiArrowLeft, FiUser } from 'react-icons/fi';
 import { useNavigate } from 'react-router-dom';
-import theme from '../styles/theme';
-import { UserManagement } from '../pages/Login/contexts/UserManagement';
+import theme from '../../../styles/theme';
+import { UserManagement } from '../../Login/contexts/UserManagement';
 
 const NavBar = () => {
   const navigate = useNavigate();
